[front/components/assistant/conversation/input_bar] - refactor: streamline input bar attachment components
 - Consolidate file and node attachment handling into a single InputBarAttachments component
 - Remove InputBarNodeAttachments and InputBarCitations components in favor of unified attachment system
 - Introduce InputBarAttachmentsPicker for selecting and uploading attachments with search capabilities

[front/components/assistant/conversation/input_bar/InputBarAttachmentsPicker] - feature: add input bar attachments picker for files and nodes

 - Implement a dropdown menu with search functionality to attach files or nodes to the input bar
 - Enable file uploading directly from the input bar with a new UI component for better user experience
