+++ front/components/assistant/conversation/input_bar/InputBar.tsx
@@ -13,19 +13,18 @@ import { useContext, useEffect, useMemo, useRef, useState } from "react";
 
 import { useFileDrop } from "@app/components/assistant/conversation/FileUploaderContext";
 import { GenerationContext } from "@app/components/assistant/conversation/GenerationContextProvider";
-import { InputBarCitations } from "@app/components/assistant/conversation/input_bar/InputBarCitations";
 import type { InputBarContainerProps } from "@app/components/assistant/conversation/input_bar/InputBarContainer";
 import InputBarContainer, {
   INPUT_BAR_ACTIONS,
 } from "@app/components/assistant/conversation/input_bar/InputBarContainer";
 import { InputBarContext } from "@app/components/assistant/conversation/input_bar/InputBarContext";
-import { InputBarNodeAttachments } from "@app/components/assistant/conversation/input_bar/InputBarNodeAttachments";
 import { useFileUploaderService } from "@app/hooks/useFileUploaderService";
 import type { DustError } from "@app/lib/error";
 import { useUnifiedAgentConfigurations } from "@app/lib/swr/assistants";
 import { useConversation } from "@app/lib/swr/conversations";
 import { useSpaces } from "@app/lib/swr/spaces";
 import { classNames } from "@app/lib/utils";
+import { InputBarAttachments } from "@app/components/assistant/conversation/input_bar/InputBarAttachments";
 
 const DEFAULT_INPUT_BAR_ACTIONS = [...INPUT_BAR_ACTIONS];
 
@@ -326,12 +325,14 @@ export function AssistantInputBar({
             )}
           >
             <div className="relative flex w-full flex-1 flex-col">
-              <InputBarNodeAttachments
-                nodes={attachedNodes}
-                spacesMap={spacesMap}
-                onRemoveNode={handleNodesAttachmentRemove}
+              <InputBarAttachments
+                files={{ service: fileUploaderService }}
+                nodes={{
+                  items: attachedNodes,
+                  spacesMap,
+                  onRemove: handleNodesAttachmentRemove,
+                }}
               />
-              <InputBarCitations fileUploaderService={fileUploaderService} />
               <InputBarContainer
                 actions={actions}
                 disableAutoFocus={disableAutoFocus}
